tests/requirements.txt
pytest>=7.4.0
pytest-mock>=3.11.1
requests>=2.31.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
